Try breaking the controls up into separate groups to allow three different breakpoint arrangements
