+++ packages/media-editor/src/components/media-editor-modal/style.scss
@@ -21,52 +21,80 @@
 	min-height: 0;
 }
 
-// Footer chrome. Default layout stacks (narrow); `.is-wide` flips to a
-// single row. Wide and narrow are driven by *different JSX trees* selected
-// via `useViewportMatch( 'xlarge' )` in the media editor so DOM order always
-// matches visual order — no `order` reshuffling, which the project's
-// stylelint rule treats as an a11y red flag.
+// Footer chrome — three layouts driven by `is-wide` / `is-medium` /
+// `is-narrow` modifier classes set from the media editor. Wide is a single
+// flex row; medium and narrow stack via `flex-direction: column`. Each
+// layout has its own JSX tree (not `order` reshuffling, which the project's
+// stylelint rule treats as an a11y red flag), so DOM order always matches
+// visual order.
 .media-editor-modal.components-modal__frame .media-editor-modal__footer {
 	flex-shrink: 0;
 	display: flex;
-	flex-direction: column;
 	gap: $grid-unit-20;
 	padding: $grid-unit-10 $grid-unit-30;
 	border-top: $border-width solid $gray-300;
 	background: $white;
 
-	// Toolbar's intrinsic padding is suppressed in favor of the footer's
-	// padding rhythm in both layouts.
-	.media-editor-toolbar {
-		padding: 0;
-	}
-
 	&.is-wide {
 		flex-direction: row;
 		align-items: center;
 
-		// Toolbar fills the middle of the single row.
-		.media-editor-toolbar {
+		// Wide-only wrapper: ruler + transform controls are grouped
+		// together, centered between History and Cancel/Save with a
+		// divider rule between them — preserving the long-standing
+		// wide-desktop look.
+		.media-editor-modal__footer-toolbar {
 			flex: 1 1 0;
 			min-width: 0;
+			display: flex;
+			align-items: center;
+			justify-content: center;
+			gap: $grid-unit-10;
 		}
 
-		// `margin-left: auto` guarantees right alignment whether or not a
-		// toolbar / history sibling is rendered (handles non-image
+		.media-editor-fine-rotation {
+			flex: 1 1 280px;
+			min-width: 0;
+			max-width: 360px;
+		}
+
+		.media-editor-transform-controls {
+			border-left: $border-width solid $gray-300;
+			padding-left: $grid-unit-15;
+		}
+
+		// `margin-left: auto` guarantees right alignment whether or not
+		// History / the toolbar is rendered (handles non-image
 		// attachments where only `FooterActions` is present).
 		.media-editor__footer-actions {
 			margin-left: auto;
 		}
 	}
+
+	&.is-medium,
+	&.is-narrow {
+		flex-direction: column;
+	}
+
+	// Narrow only: transform cluster sits on its own row, centered.
+	&.is-narrow > .media-editor-transform-controls {
+		align-self: center;
+	}
 }
 
-// Narrow-only secondary row: groups history + Cancel/Save onto one line
-// beneath the toolbar.
+// Secondary row used by medium (History + Transform + Cancel/Save) and
+// narrow (History + Cancel/Save). `margin-left: auto` on the transform
+// cluster and on Cancel/Save splits any free space — History gets the
+// left edge, Cancel/Save the right edge, and the transform cluster (when
+// present, i.e. medium only) approximately centers. Exact centering only
+// holds when History and Cancel/Save have matching widths; in practice
+// they're close enough that the offset is imperceptible.
 .media-editor-modal__footer-row {
 	display: flex;
 	align-items: center;
 	gap: $grid-unit-20;
 
+	.media-editor-transform-controls,
 	.media-editor__footer-actions {
 		margin-left: auto;
 	}
