+++ packages/media-editor/src/style.scss
@@ -4,6 +4,6 @@
 @use "./components/media-editor/style.scss" as *;
 @use "./components/media-editor-modal/style.scss" as *;
 @use "./components/media-editor-canvas/style.scss" as *;
-@use "./components/media-editor-toolbar/style.scss" as *;
+@use "./components/media-editor-transform-controls/style.scss" as *;
 @use "./components/media-editor-keyboard-shortcuts-modal/style.scss" as *;
 @use "./components/rotation-ruler/style.scss" as *;
